add command detect_trail

+++ src/opencv/opencv.h
@@ -30,7 +30,7 @@ int cvUnsharpFilter(fits*, double, double);
 
 int cvClahe(fits *image, double clip_limit, int size);
 
-int cvHoughLines(fits *image, int layer, int threshvalue, int minlen);
+int cvHoughLines(fits *image, int layer, float threshvalue, int minlen);
 
 #ifdef __cplusplus
 }
